[Update+Create]Refactored Playermovement Managementand updated directories Movement Scripts and seperated camera logic and calculations
